backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
azure-identity>=1.16.0
openapi-spec-validator>=0.7.1
pydantic>=2.0.0
pytest>=8.0.0
